+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Move `Protocol::URL::PATTERN` to `protocol/url/pattern.rb` so it can be shared more easily.
+
 ## v0.1.0
 
   - Initial implementation.